*** empty log message ***
git-svn-id: https://crosswire.org/svn/sword/trunk@1361 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ src/modules/common/swcipher.cpp
@@ -70,7 +70,7 @@ char *SWCipher::cipherBuf(unsigned long *ilen, const char *ibuf)
 
 	Encode();
 
-	*ilen = (short)len;
+	*ilen = len;
 	return buf;
 }
